requirements.txt
Pillow==9.0.1
PyQt5==5.15.6